wings_scraper/requirements.txt
playwright==1.44.0
supabase==2.4.6
python-dotenv==1.0.1
